CR129 Moved DTO object to more appropriate package. Minor fix to unit test
